flask-ai-copilot/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
openai>=1.40.0
google-generativeai==0.3.2
groq>=0.4.1

